Fucking react and its need for performance tuning what is this the nineties?
