--- src/components/bytecodeDebugger/ContextSearchList.module.scss
@@ -1,1 +1,0 @@
-@import "../../theme";
