+++ src/components/bytecodeDebugger/Id.module.scss
@@ -1,6 +1,6 @@
 @import "../../theme";
 
-.id {
+.container {
   color: $color-font-link;
   font-style: italic;
   cursor: pointer;
