--- src/components/bytecodeDebugger/InstanceSearchList.module.scss
@@ -1,1 +1,0 @@
-@import "../../theme";
